refactor: extract cell utils to new files
