+++ src/components/Pane.tsx
@@ -2,7 +2,8 @@ import React, { CSSProperties } from "react";
 import { NumericalRange } from "../types/CellMatrix";
 import { GetCellOffsets, PaneName } from "../types/InternalModel";
 import { RGTheme } from "../types/Theme";
-import { areAreasEqual, isSpanMember } from "../utils/cellUtils";
+import { isSpanMember } from "../utils/isSpanMember";
+import { areAreasEqual } from "../utils/areAreasEqual";
 import { useReactGridStore } from "../utils/reactGridStore";
 import { useTheme } from "../hooks/useTheme";
 import { CellContext } from "./CellContext";
