+++ src/utils/getCellFromCertainPane.ts
@@ -1,5 +1,5 @@
 import { PaneName } from "../types/InternalModel";
-import { isCellInRange } from "./cellUtils";
+import { isCellInRange } from "./isCellInRange";
 import { getCellIndexesFromContainerElement } from "./getCellIndexes";
 import { ReactGridStore } from "./reactGridStore";
