+++ src/utils/isSpanMember.ts
@@ -1,0 +1,6 @@
+import { Cell, SpanMember } from "../types/PublicModel";
+
+
+export const isSpanMember = (cell: Cell | SpanMember): cell is SpanMember => {
+  return "originRowId" in cell && "originColId" in cell;
+};
